@@ -1,6 +1,6 @@
-#region Copyright (C) 2017-2021 Yaroslav Tatarenko
+#region Copyright (C) 2017-2022 Yaroslav Tatarenko
 
-// Copyright (C) 2017-2021 Yaroslav Tatarenko
+// Copyright (C) 2017-2022 Yaroslav Tatarenko
 // This product uses MediaInfo library, Copyright (c) 2002-2021 MediaArea.net SARL. 
 // https://mediaarea.net
 
@@ -34,13 +34,5 @@ namespace ApiSample.Infrastructure
         public static IServiceCollection AddFilters(this IServiceCollection services) =>
             services
                 .AddScoped<ValidateModelStateAttribute>();
-
-        /// <summary>
-        /// Adds the MediaInfo logger to DI container.
-        /// </summary>
-        /// <param name="services">The services instance.</param>
-        public static IServiceCollection AddMediaInfoLogger(this IServiceCollection services) =>
-            services
-                .AddScoped<MediaInfo.ILogger, MediaInfoLogger>();
     }
 }
