@@ -1,6 +1,6 @@
-﻿#region Copyright (C) 2017-2021 Yaroslav Tatarenko
+﻿#region Copyright (C) 2017-2022 Yaroslav Tatarenko
 
-// Copyright (C) 2017-2021 Yaroslav Tatarenko
+// Copyright (C) 2017-2022 Yaroslav Tatarenko
 // This product uses MediaInfo library, Copyright (c) 2002-2021 MediaArea.net SARL. 
 // https://mediaarea.net
 
@@ -239,6 +239,7 @@ namespace ApiSample.Models
         /// <example>1992-03-06</example>
         [DataMember(Name = "writtenDate")]
         [JsonPropertyName("writtenDate")]
+        [JsonConverter(typeof(JsonMicrosoftDateTimeConverter))]
         public DateTime? WrittenDate { get; set; }
 
         /// <summary>
