@@ -1,6 +1,6 @@
-#region Copyright (C) 2017-2021 Yaroslav Tatarenko
+#region Copyright (C) 2017-2022 Yaroslav Tatarenko
 
-// Copyright (C) 2017-2021 Yaroslav Tatarenko
+// Copyright (C) 2017-2022 Yaroslav Tatarenko
 // This product uses MediaInfo library, Copyright (c) 2002-2021 MediaArea.net SARL. 
 // https://mediaarea.net
 
@@ -49,7 +49,6 @@ namespace ApiSample
 
             services
                 .AddFilters()
-                .AddMediaInfoLogger()
                 .AddAutoMapper(MapperExtensions.ConfigureMapper)
                 .AddSwaggerGen(options =>
                 {
